gestion de la fonction profile
Supressin du controller manageprofile , pour fusionner la fonction profile dans le controller user
